api/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
llama-index==0.6.34
pandas==2.0.2
openai==0.27.8
pyjwt==2.7.0
openpyxl==3.1.2